Moved around examples
