+++ example/alpha/internal/excluded/ok.go
@@ -1,0 +1,7 @@
+package excluded
+
+import "github.com/TheFellow/go-arch-lint/example/alpha/experimental"
+
+func Test() {
+	_ = experimental.NewWidget()
+}
