+++ example/alpha/tester.go
@@ -1,0 +1,10 @@
+package alpha
+
+import (
+	"github.com/TheFellow/go-arch-lint/example/alpha/internal/exception"
+)
+
+func DoTest() {
+	exception.Test()
+	//_ = experimental.NewWidget()
+}
